*Now containers are stored properly as well when using binary map storage. (I thought the functions to store these were already defined, I was mistaken. :|)

+++ otserv/trunk/fileloader.h
@@ -272,8 +272,7 @@ public:
 		return true;
 	}
 
-
-	inline bool SKIP_N(unsigned short n){
+	inline bool SKIP_N(int32_t n){
 		if(size() < n){
 			return false;
 		}
